notificationsInit - move API->internal conversion to a shared function
and prefer map over forEach+push

shared function because the same code exists in the addNotification action
