+++ app/javascript/packs/notification-drawer.js
@@ -2,38 +2,27 @@ import { get } from 'lodash';
 import { listenToRx } from '../miq_observable';
 import { API } from '../http_api';
 import { addNotification, initNotifications } from '../miq-redux/actions/notifications-actions';
-import { maxNotifications } from '../notifications/backend.js';
+import { maxNotifications, convert } from '../notifications/backend.js';
 
 export function notificationsInit(useLimit) {
-  const notifications = [];
   const promises = [];
   const limitFragment = useLimit ? `&limit=${maxNotifications}` : '';
+
   promises.push(API.get(`/api/notifications?expand=resources&attributes=details&sort_by=id&sort_order=desc${limitFragment}`)
     .then((data) => {
-      data.resources.forEach((resource) => {
-        const message = window.miqFormatNotification(resource.details.text, resource.details.bindings);
-        notifications.push({
-          id: resource.id,
-          notificationType: 'event',
-          unread: !resource.seen,
-          type: resource.details.level,
-          message,
-          data: {
-            link: get(resource.details, 'bindings.link'),
-          },
-          href: resource.href,
-          timeStamp: resource.details.created_at,
-        });
-      });
+      const notifications = data.resources.map(convert);
+
       return {
         notifications,
         subcount: data.subcount,
       };
     }));
+
   if (useLimit) {
     // get real subcount
     promises.push(API.get('/api/notifications'));
   }
+
   return Promise.all(promises).then(([{ notifications, subcount }, meta]) => ({
     notifications,
     subcount: meta ? meta.subcount : subcount,
